Handle missing mountpoints, support pool restrictions
Say you want to snapshot only `tank`, excluding `rpool`, this will
likely leave you without a / on which to mount filesystems from `tank`.

Similarly if you're not on a ZFS-on-root system, ZFS is likely rooted
elsewhere than /, which won't exist on the target.

Handle these cases by searching up the tree for a mountpoint that *does*
exist, mounting a tmpfs on it, and then mkdir_p the mountpoint we need.

Change mountpoint recording to add them as they're created rather than
building the mountpoint list twice.

Add a -p / --pool flag to specify only the pools to involve in zfsnapr
operation.

Bump to 0.2.

This *should* resolve #2.